scrapping-buscacursos/pyproject.toml
[tool.poetry]
name = "scrapping-buscacursos"
version = "0.1.0"
description = ""
authors = ["Hernan Valdivieso <hfvaldivieso@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
pandas = "^1.2.0"
python-dotenv = "^0.15.0"
pymongo = "^3.11.2"
dnspython = "^2.1.0"
black = "^20.8b1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
